pyproject.toml
[project]
name = "tally-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "celery>=5.6.2",
    "cryptography>=46.0.5",
    "defusedxml>=0.7.1",
    "lxml>=6.0.2",
    "marshmallow>=4.2.1",
    "openpyxl>=3.1.5",
    "pandas>=3.0.0",
    "pydantic>=2.12.5",
    "pymysql>=1.1.2",
    "python-dotenv==1.0.0",
    "redis>=7.1.0",
    "requests>=2.32.5",
    "rq>=2.6.1",
    "sqlalchemy>=2.0.46",
    "tabulate>=0.9.0",
    "xmltodict>=1.0.2",
]
